Move alien-signals code to system.ts
