+++ packages/reactivity/__tests__/effect.spec.ts
@@ -8,14 +8,7 @@ import {
   serializeInner,
 } from '@vue/runtime-test'
 import { ITERATE_KEY, getDepFromReactive } from '../src/dep'
-import {
-  type Dependency,
-  endBatch,
-  onEffectCleanup,
-  pauseTracking,
-  resetTracking,
-  startBatch,
-} from '../src/effect'
+import { onEffectCleanup, pauseTracking, resetTracking } from '../src/effect'
 import {
   type DebuggerEvent,
   type ReactiveEffectRunner,
@@ -29,6 +22,7 @@ import {
   stop,
   toRaw,
 } from '../src/index'
+import { Dependency, endBatch, startBatch } from '../src/system'
 
 describe('reactivity/effect', () => {
   it('should run the passed function once (wrapped by a effect)', () => {
