+++ packages/reactivity/src/arrayInstrumentations.ts
@@ -1,8 +1,9 @@
 import { isArray } from '@vue/shared'
 import { TrackOpTypes } from './constants'
 import { ARRAY_ITERATE_KEY, track } from './dep'
-import { endBatch, pauseTracking, resetTracking, startBatch } from './effect'
+import { pauseTracking, resetTracking } from './effect'
 import { isProxy, isShallow, toRaw, toReactive } from './reactive'
+import { endBatch, startBatch } from './system'
 
 /**
  * Track array iteration and return:
