+++ packages/reactivity/src/dep.ts
@@ -1,16 +1,15 @@
 import { isArray, isIntegerKey, isMap, isSymbol } from '@vue/shared'
 import { type TrackOpTypes, TriggerOpTypes } from './constants'
 import { onTrack, triggerEventInfos } from './debug'
+import { activeSub, activeTrackId } from './effect'
 import {
   type Dependency,
   type Link,
-  activeSub,
-  activeTrackId,
   endBatch,
   link,
   propagate,
   startBatch,
-} from './effect'
+} from './system'
 
 class Dep implements Dependency {
   _subs: Link | undefined = undefined
